+++ R/PAMcorrection-Package.R
@@ -3,7 +3,7 @@
 #' @name PAMcorrection - package
 #' @docType package
 #' @title PAMcorrection - a tool for PAM correction
-#' @description The \code{PAMcorrection} package is used to estimate and test correction vlaues for the PAM using control values.
+#' @description The \code{PAMcorrection} package is used to estimate and test correction values for the PAM using control values.
 #' @author Andreas Schönberg
 #' @keywords package
 NULL